+++ proxmoxtf/data_source_virtual_environment_dns_test.go
@@ -1,0 +1,43 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+package proxmoxtf
+
+import (
+	"github.com/hashicorp/terraform/helper/schema"
+	"testing"
+)
+
+// TestDataSourceVirtualEnvironmentDNSInstantiation tests whether the DataSourceVirtualEnvironmentDNS instance can be instantiated.
+func TestDataSourceVirtualEnvironmentDNSInstantiation(t *testing.T) {
+	s := dataSourceVirtualEnvironmentDNS()
+
+	if s == nil {
+		t.Fatalf("Cannot instantiate dataSourceVirtualEnvironmentDNS")
+	}
+}
+
+// TestDataSourceVirtualEnvironmentDNSSchema tests the dataSourceVirtualEnvironmentDNS schema.
+func TestDataSourceVirtualEnvironmentDNSSchema(t *testing.T) {
+	s := dataSourceVirtualEnvironmentDNS()
+
+	testRequiredArguments(t, s, []string{
+		mkDataSourceVirtualEnvironmentDNSNodeName,
+	})
+
+	testComputedAttributes(t, s, []string{
+		mkDataSourceVirtualEnvironmentDNSDomain,
+		mkDataSourceVirtualEnvironmentDNSServers,
+	})
+
+	testSchemaValueTypes(t, s, []string{
+		mkDataSourceVirtualEnvironmentDNSDomain,
+		mkDataSourceVirtualEnvironmentDNSNodeName,
+		mkDataSourceVirtualEnvironmentDNSServers,
+	}, []schema.ValueType{
+		schema.TypeString,
+		schema.TypeString,
+		schema.TypeList,
+	})
+}