+++ proxmoxtf/provider.go
@@ -31,6 +31,7 @@ func Provider() *schema.Provider {
 		ConfigureFunc: providerConfigure,
 		DataSourcesMap: map[string]*schema.Resource{
 			"proxmox_virtual_environment_datastores": dataSourceVirtualEnvironmentDatastores(),
+			"proxmox_virtual_environment_dns":        dataSourceVirtualEnvironmentDNS(),
 			"proxmox_virtual_environment_group":      dataSourceVirtualEnvironmentGroup(),
 			"proxmox_virtual_environment_groups":     dataSourceVirtualEnvironmentGroups(),
 			"proxmox_virtual_environment_nodes":      dataSourceVirtualEnvironmentNodes(),
@@ -43,6 +44,7 @@ func Provider() *schema.Provider {
 			"proxmox_virtual_environment_version":    dataSourceVirtualEnvironmentVersion(),
 		},
 		ResourcesMap: map[string]*schema.Resource{
+			"proxmox_virtual_environment_dns":   resourceVirtualEnvironmentDNS(),
 			"proxmox_virtual_environment_file":  resourceVirtualEnvironmentFile(),
 			"proxmox_virtual_environment_group": resourceVirtualEnvironmentGroup(),
 			"proxmox_virtual_environment_pool":  resourceVirtualEnvironmentPool(),